+++ python/mrc/_pymrc/tests/test_pipeline.cpp
@@ -122,7 +122,7 @@ TEST_F(TestPipeline, Execute)
         });
 
         auto internal = seg.make_node<bool, unsigned int>("internal", rxcpp::operators::map([](bool b) {
-                                                              unsigned int i{b};
+                                                              unsigned int i{static_cast<unsigned int>(b)};
                                                               return i;
                                                           }));
